+++ src/cz/mg/application/entities/buildin/atoms/float32/MgFloat32BinaryOperatorDivide.java
@@ -17,7 +17,7 @@ public class MgFloat32BinaryOperatorDivide extends MgFloat32BinaryOperator {
 
 
     @Override
-    protected int compute(int left, int right) {
+    protected float compute(float left, float right) {
         return left / right;
     }
 }
